WIP menu updates, kingdom menu moved #285
