+++ core/src/main/java/com/github/rumsfield/konquest/display/icon/MenuIcon.java
@@ -1,13 +1,18 @@
 package com.github.rumsfield.konquest.display.icon;
 
+import com.github.rumsfield.konquest.display.StateMenu;
 import org.bukkit.inventory.ItemStack;
+
+import javax.annotation.Nullable;
 
 public abstract class MenuIcon {
 
-	private int index;
+	private int index; // The slot index in the inventory of this icon
+	private StateMenu.State state; // The state that this icon will update the menu to
 
 	public MenuIcon(int index) {
 		this.index = index;
+		this.state = null;
 	}
 
 	public int getIndex() {
@@ -16,6 +21,14 @@ public abstract class MenuIcon {
 
 	public void setIndex(int val) {
 		this.index = val;
+	}
+
+	public @Nullable StateMenu.State getState() {
+		return state;
+	}
+
+	public void setState(StateMenu.State state) {
+		this.state = state;
 	}
 
 	public abstract String getName();
