+++ core/src/main/java/com/github/rumsfield/konquest/utility/MessagePath.java
@@ -29,6 +29,7 @@ public enum MessagePath {
 	LABEL_FLAGS                                 (0, "label.flags"),
 	LABEL_GOLEM_OFFENSE                         (0, "label.golem-offense"),
 	LABEL_HEALTH                                (0, "label.health"),
+	LABEL_HOME                                  (0, "label.home"),
 	LABEL_IMMUNITY                              (0, "label.immunity"),
 	LABEL_INCREMENT_COST                        (0, "label.increment-cost"),
 	LABEL_INFORMATION                          	(0, "label.information"),
@@ -71,6 +72,7 @@ public enum MessagePath {
 	LABEL_RESIDENCIES                           (0, "label.residencies"),
 	LABEL_RESIDENT                             	(0, "label.resident"),
 	LABEL_RESIDENTS                             (0, "label.residents"),
+	LABEL_RETURN                                (0, "label.return"),
 	LABEL_RUIN                                  (0, "label.ruin"),
 	LABEL_SANCTUARY                             (0, "label.sanctuary"),
 	LABEL_SCORE                                 (0, "label.score"),
