+++ core/src/main/resources/lang/english.yml
@@ -35,6 +35,7 @@ label:
   flags: "Flags"
   golem-offense: "Iron Golem Offense"
   health: "Health"
+  home: "Home"
   immunity: "Immunity"
   increment-cost: "Increment Cost"
   information: "Information"
@@ -77,6 +78,7 @@ label:
   residencies: "Residencies"
   resident: "Resident"
   residents: "Residents"
+  return: "Return"
   ruin: "Ruin"
   sanctuary: "Sanctuary"
   score: "Score"
